Support list of private constants (#19)
* restore tests from #5

* update validator to support private_constants setting

* add private_constants property

With this change, the `enforce_privacy` setting will support only four
values: true, false, 'strict', and nil. The array of private constant
names can now only be specified via a new setting
called `private_constants`. If `enforce_privacy` is set to false or
nil, these private constants are ignored by the checker. If
`private_constants` is not specified or is an empty list, all
constants in the package are assumed to be private unless
they are defined in the public directory.

* bump minor version